Revert "Endereço DAO e Endereço Serviço Adicionados"
This reverts commit 6be14feda6a1992c8ee6332b875aff0eb8c3ee90.

+++ src/dao/ClienteDAO.java
@@ -11,7 +11,6 @@ import utils.Conexao;
 public class ClienteDAO implements IGerenciamentoDAO {
 
 	private Cliente cliente;
-	
 	private Conexao conexao;
 	
 	public Cliente getCliente() {
@@ -103,7 +102,7 @@ public class ClienteDAO implements IGerenciamentoDAO {
 			while(result.next()) {
 				
 				Cliente cliente = new Cliente();
-				cliente.setId(result.getLong("id"));
+				cliente.setId(result.getInt("id"));
 				cliente.setNomeCliente(result.getString("nome"));
 				
 				clientes.add(cliente);
@@ -135,7 +134,7 @@ public class ClienteDAO implements IGerenciamentoDAO {
 			while(result.next()) {
 				
 				Cliente cliente = new Cliente();
-				cliente.setId(result.getLong("id"));
+				cliente.setId(result.getInt("id"));
 				cliente.setNomeCliente(result.getString("nome"));
 				
 				clientes.add(cliente);
